Fixed conflicts
- Fixed Actor and Movie classes for conflict

+++ moovi/src/main/java/com/dbms/project/moovi/data/repository/MovieRepository.java
@@ -3,5 +3,5 @@ package com.dbms.project.moovi.data.repository;
 import com.dbms.project.moovi.data.entity.Movie;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MovieRepository extends CrudRepository<Movie, Integer>{
+public interface MovieRepository extends CrudRepository<Movie, Long>{
 }
